fix: 更新 FingerprintsBody 和 FingerprintResult 结构体中的 fingerprints 字段类型为 i64
[2025-06-20T02:12:52Z DEBUG actix_web::middleware::logger] Error in response: BadRequest("Json deserialize error: invalid value: integer `2413029736`, expected i32 at line 1 column 27")

+++ src/services/curseforge.rs
@@ -392,7 +392,7 @@ impl CurseforgeService {
 
     pub async fn get_fingerprints(
         &self,
-        fingerprints: Vec<i32>,
+        fingerprints: Vec<i64>,
         game_id: Option<i32>,
     ) -> Result<FingerprintResponse, ServiceError> {
         if fingerprints.is_empty() {
